Move docs folder to features folder
* Extract docs specific utils and constants
