+++ packages/ag-charts-website/src/astro/integrations/agFileWatcher.ts
@@ -1,5 +1,5 @@
 import { type AstroIntegration } from 'astro';
-import { getAllDocsExampleFileList } from '../../utils/pages';
+import { getAllExamplesFileList } from '../../features/docs/utils/filesData';
 
 /**
  * Astro plugin to watch AG files, so dev server restarts when they are changed
@@ -13,7 +13,7 @@ export default function createAgFileWatcherIntegration(): AstroIntegration {
                     return;
                 }
 
-                const exampleFiles = await getAllDocsExampleFileList();
+                const exampleFiles = await getAllExamplesFileList();
                 exampleFiles.forEach((exampleFile) => {
                     addWatchFile(exampleFile);
                 });
