+++ packages/ag-charts-website/src/features/docs/components/DocsExampleRunner.astro
@@ -1,9 +1,9 @@
 ---
 import { DocsExampleRunner } from './DocsExampleRunner';
-import { getExampleOptionsFromString } from '../../features/example-runner/utils/getExampleOptionsFromString';
-import { DOCS_FRAMEWORK_PATH_INDEX, DOCS_PAGE_NAME_PATH_INDEX } from '../../constants';
-import type { ExampleType } from '../../features/examples-generator/types';
-import ExampleRunnerContainer from '../../features/example-runner/components/ExampleRunnerContainer.astro';
+import { getExampleOptionsFromString } from '../../example-runner/utils/getExampleOptionsFromString';
+import type { ExampleType } from '../../examples-generator/types';
+import ExampleRunnerContainer from '../../example-runner/components/ExampleRunnerContainer.astro';
+import { DOCS_FRAMEWORK_PATH_INDEX, DOCS_PAGE_NAME_PATH_INDEX } from '../constants';
 
 interface Props {
     title: string;
