+++ packages/ag-charts-website/src/features/docs/components/Header.module.scss
@@ -1,4 +1,4 @@
-@use '../../design-system' as *;
+@use '../../../design-system' as *;
 
 .docsPageHeader {
     position: relative;
