+++ packages/ag-charts-website/src/features/docs/components/Header.tsx
@@ -1,8 +1,8 @@
 import type { FunctionComponent } from 'react';
 import styles from './Header.module.scss';
-import { Icon } from '../icon/Icon';
-import type { Framework } from '../../types/ag-grid';
-import { getFrameworkDisplayText } from '../../utils/framework';
+import { Icon } from '../../../components/icon/Icon';
+import type { Framework } from '../../../types/ag-grid';
+import { getFrameworkDisplayText } from '../../../utils/framework';
 
 interface Props {
     title: string;
