+++ packages/ag-charts-website/src/features/docs/constants.ts
@@ -1,0 +1,4 @@
+import { SITE_BASE_URL_SEGMENTS } from '../../constants';
+
+export const DOCS_FRAMEWORK_PATH_INDEX = SITE_BASE_URL_SEGMENTS + 1;
+export const DOCS_PAGE_NAME_PATH_INDEX = SITE_BASE_URL_SEGMENTS + 2;
