+++ packages/ag-charts-website/src/layouts/Layout.astro
@@ -6,7 +6,7 @@ import '../design-system/design-system.scss';
 import { Footer } from '../components/footer/Footer';
 import { TopBar } from '../components/top-bar/TopBar';
 import { FRAMEWORKS } from '../constants';
-import { getFrameworkFromPath } from '../utils/pages';
+import { getFrameworkFromPath } from '../features/docs/utils/urlPaths';
 
 export interface Props {
     title?: string;
