+++ packages/ag-charts-website/src/pages/[framework]/[pageName].astro
@@ -1,11 +1,12 @@
 ---
 import { getCollection } from 'astro:content';
 import Layout from '../../layouts/Layout.astro';
-import { getDocsPages, getDocsPagesList } from '../../utils/pages';
 import { PagesNavigation } from '../../components/pages-navigation/PagesNavigation';
-import { Header } from '../../components/docs/Header';
+import { Header } from '../../features/docs/components/Header';
 import { getFrameworkDisplayText } from '../../utils/framework';
 import type { Framework } from '../../types/ag-grid';
+import { getPagesList } from '../../features/docs/utils/filesData';
+import { getDocsPages } from '../../features/docs/utils/pageData';
 
 export async function getStaticPaths() {
     const pages = await getCollection('docs');
@@ -19,7 +20,7 @@ const { title, enterprise, sideMenu = true } = page.data;
 const { Content } = await page.render();
 
 const allDocPages = await getCollection('docs');
-const pages = getDocsPagesList(allDocPages);
+const pages = getPagesList(allDocPages);
 ---
 
 <Layout title={`${getFrameworkDisplayText(framework as Framework)} Charts: AG Charts`} showTopBar={true}>
