+++ packages/ag-charts-website/src/pages/[internalFramework]/[pageName]/examples/[exampleName].astro
@@ -1,12 +1,12 @@
 ---
 import { getCollection } from 'astro:content';
-import { getDocExamplePages } from '../../../../utils/pages';
 import type { InternalFramework } from '../../../../types/ag-grid';
-import DocsFrameworkTemplate from '../../../../components/docs/DocsFrameworkTemplate.astro';
+import DocsFrameworkTemplate from '../../../../features/docs/components/DocsFrameworkTemplate.astro';
+import { getDocsExamplePages } from '../../../../features/docs/utils/pageData';
 
 export async function getStaticPaths() {
     const pages = await getCollection('docs');
-    return getDocExamplePages({
+    return getDocsExamplePages({
         pages,
     });
 }
