+++ packages/ag-charts-website/src/pages/[internalFramework]/[pageName]/examples/[exampleName]/[fileName].ts
@@ -1,8 +1,9 @@
 import { getCollection } from 'astro:content';
-import { getDocExampleFiles, getContentRootFileUrl } from '../../../../../utils/pages';
+import { getContentRootFileUrl } from '../../../../../utils/pages';
 import { getIsDev } from '../../../../../utils/env';
-import { getGeneratedDocsContents } from '../../../../../features/examples-generator/examplesGenerator';
+import { getGeneratedDocsContents } from '../../../../../features/docs/utils/examplesGenerator';
 import type { InternalFramework } from '../../../../../types/ag-grid';
+import { getDocExampleFiles } from '../../../../../features/docs/utils/pageData';
 
 interface Params {
     internalFramework: InternalFramework;
