+++ packages/ag-charts-website/src/pages/debug/docs-example-files.astro
@@ -1,17 +1,18 @@
 ---
 import { getCollection } from 'astro:content';
-import { getDocExamplePages, getDocsExampleUrl } from '../../utils/pages';
 import Layout from '../../layouts/Layout.astro';
+import { getDocExampleFiles } from '../../features/docs/utils/pageData';
+import { getExampleFileUrl } from '../../features/docs/utils/urlPaths';
 
 const pages = await getCollection('docs');
-const examples = await getDocExamplePages({
+const exampleFiles = await getDocExampleFiles({
     pages,
 });
 ---
 
-<Layout title={`Debug: Example Pages`}>
+<Layout title={`Debug: Docs Example Files`}>
     <main class="page-margin container">
-        <h1>Example pages ({examples.length})</h1>
+        <h1>Example files ({exampleFiles.length})</h1>
         <table>
             <thead>
                 <th>No.</th>
@@ -19,15 +20,17 @@ const examples = await getDocExamplePages({
                 <th><code>params.internalFramework</code></th>
                 <th><code>params.pageName</code></th>
                 <th><code>params.exampleName</code></th>
+                <th><code>params.fileName</code></th>
             </thead>
             <tbody>
                 {
-                    examples.map((example, index) => {
-                        const { internalFramework, pageName, exampleName } = example.params;
-                        const url = getDocsExampleUrl({
+                    exampleFiles.map((example, index) => {
+                        const { internalFramework, pageName, exampleName, fileName } = example.params;
+                        const url = getExampleFileUrl({
                             internalFramework,
                             pageName,
                             exampleName,
+                            fileName,
                         });
                         return (
                             <tr>
@@ -38,6 +41,7 @@ const examples = await getDocExamplePages({
                                 <td>{internalFramework}</td>
                                 <td>{pageName}</td>
                                 <td>{exampleName}</td>
+                                <td>{fileName}</td>
                             </tr>
                         );
                     })
